Change some files location
